Fixed jacob contest time shows one second too long

+++ src/main/java/at/hannibal2/skyhanni/features/garden/contest/JacobContestStatsSummary.kt
@@ -39,7 +39,7 @@ class JacobContestStatsSummary {
             }
             FarmingContestPhase.STOP -> {
                 val duration = System.currentTimeMillis() - startTime
-                val durationInSeconds = duration / 1000
+                val durationInSeconds = duration.toDouble() / 1000
                 val blocksPerSecond = (blocksBroken.toDouble() / durationInSeconds).round(2)
                 val cropName = event.crop.cropName
                 LorenzUtils.chat("§e[SkyHanni] Stats for $cropName Contest:")
